requirements.txt
cycler==0.10.0
dill==0.2.9
futures==3.1.1
kiwisolver==1.0.1
matplotlib==3.0.2
multiprocess==0.70.7
numpy==1.16.1
Pillow==5.4.1
pyparsing==2.3.1
python-dateutil==2.8.0
six==1.12.0
